refactor: move the helper function to a dedicated trait
